main_project1/requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
osmnx==1.3.0
networkx==2.8.4
geopandas==0.10.2
shapely==2.0.1
folium==0.12.1